feat: delete 비ë°밀번호 확인 추가

+++ src/main/java/org/danji/member/service/MemberServiceImpl.java
@@ -20,13 +20,6 @@ public class MemberServiceImpl implements MemberService {
 
     private final MemberMapper mapper;
 
-
-    // 아이디 중복 검사
-//    @Override
-//    public boolean checkDuplicate(String username) {
-//        return mapper.findByUsername(username) != null;
-//    }
-
     // 사용자 조회
     @Override
     public MemberDTO get(String username) {
@@ -47,13 +40,10 @@ public class MemberServiceImpl implements MemberService {
                 .role(memberJoinDTO.getRole())
                 .build();
         mapper.insert(member);
-//        AuthVO auth = new AuthVO();
-//        auth.setUsername(member.getUsername());
-//        auth.setAuth("USER");
-//        mapper.insertAuth(auth);
         return get(member.getUsername());
     }
 
+    // 정보 수정
     @Override
     public MemberDTO update(MemberUpdateDTO member) {
         MemberVO vo = mapper.get(member.getUsername());
@@ -65,9 +55,21 @@ public class MemberServiceImpl implements MemberService {
         return get(member.getUsername());
     }
 
+    // 삭제
     @Override
-    public void delete(String username) {
+    public void delete(MemberDeleteDTO dto) {
+        String username = dto.getUsername();
+        String inputPassword = dto.getPassword();
+
+        MemberVO member = mapper.get(username);
+
+        if (member == null) {
+            throw new NoSuchElementException("존재하지 않는 회원입니다.");
+        } // DB에 있는 회원인지 조회
+
+        if (!dto.getPassword().equals(member.getPassword())) {
+            throw new IllegalArgumentException("비밀번호가 일치하지 않습니다.");
+        } // 비밀번호 확인 조회
         mapper.deleteByUsername(username);
     }
-
 }
